Adição Estruturas de pastas
Adição de estruturas de pastas e extração de métodos para classes em especifico.
